requirements.txt
langchain >= 0.3.25
PyGithub >= 2.6.1
python-dotenv >= 1.1.0